Work to support ansible (#3)
* Add email format (#2)

* Add email format

This adds a custom formatter for email addresses which should get rid of
the warnings that are generated when initiating the client. Also added
support for the force_renew flag when logging in so the user can force
the creation of a new auth token.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Add tests for force_renew

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Add option to disable python models (#3)

By default bravado converts the resulting data from the API to models
that it auto generates. This change makes this behavior optional for
cases where you just want the data as a dict.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Add support for python 2.7 (#4)

* Add support for python 2.7

These are changes to make opsyclient support python 2.7.

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Fix Tests

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>

* Fix Tests2

Signed-off-by: M. David Bennett <m.david.bennett@rackspace.com>